+++ runtests.sh
@@ -468,7 +468,7 @@ function run_integration_tests() {
 # network training/inference/eval integration tests
 if [ $doNetTests = true ]; then
   run_integration_tests "radiology" "tests/data/dataset/local/spleen" "deepedit,segmentation_spleen,segmentation,deepgrow_2d,deepgrow_3d"
-  run_integration_tests "pathology" "tests/data/pathology" "segmentation_nuclei,nuclick"
+  run_integration_tests "pathology" "tests/data/pathology" "all"
   run_integration_tests "monaibundle" "tests/data/dataset/local/spleen" "spleen_ct_segmentation_v0.1.0,spleen_deepedit_annotation_v0.1.0,swin_unetr_btcv_segmentation_v0.1.0"
   run_integration_tests "endoscopy" "tests/data/endoscopy" "tooltracking,inbody,deepedit"
 fi
